pipeline/requirements.txt
trimesh>=3.20.0
numpy>=1.24.0
pyyaml>=6.0
pandas>=2.0.0
